style: remove json_parser.cpp
