+++ be/src/exec/CMakeLists.txt
@@ -94,7 +94,6 @@ set(EXEC_FILES
     vectorized/hdfs_scanner_parquet.cpp
     vectorized/hdfs_scanner_text.cpp
     vectorized/json_scanner.cpp
-    vectorized/json_parser.cpp
     vectorized/project_node.cpp
     vectorized/dict_decode_node.cpp
     vectorized/repeat_node.cpp
@@ -206,8 +205,6 @@ set(EXEC_FILES
 if ("${CMAKE_BUILD_TARGET_ARCH}" STREQUAL "x86" OR "${CMAKE_BUILD_TARGET_ARCH}" STREQUAL "x86_64")
     set_source_files_properties(vectorized/json_scanner.cpp PROPERTIES COMPILE_FLAGS -mno-avx)
     set_source_files_properties(vectorized/json_scanner.cpp PROPERTIES COMPILE_FLAGS -mno-avx2)
-    set_source_files_properties(vectorized/json_parser.cpp PROPERTIES COMPILE_FLAGS -mno-avx)
-    set_source_files_properties(vectorized/json_parser.cpp PROPERTIES COMPILE_FLAGS -mno-avx2)
 endif()
 
 set(EXEC_FILES
